Finalize the structure of generic boundary conditions for trc
Include the subroutine trcbc.F90 that implements a structure
similar to trcdta for reading surface, coastal (river) and
open boundary conditions for tracers.
A new namelist trc_bc can be included in namelist_top
where the files to be read during stepping are specified.
At the moment this structure is only used by the BFM but it
is harmless for PISCES as this model still reads those
data internally.
NOTE: OBC reading has not been implementetested as it will be made
available only for BDY.
--This line, and th se below, will be ignored--

M    trcdta.F90
A    trcbc.F90
M    par_trc.F90