Correct problem with reading mesh_mask in OFF_SRC, Ticket #1253
The reading of mesh mask information is now compliant with the
grid variables written in mesh_mask.nc.
Be aware that if one is using a mesh_mask generated with a previous 
version of NEMO (<3.6) variable e3t must be renamed to e3t_0.
This commit also adds information on the date which may be needed
by other biogeochemical models.