Special branch to help debug non bit-comparison in restarts caused by isf code.
The code itself is in sbcisf where it sets ralpha, rhisf_tbl 
and rhisf_tbl_0 on a restart in a manner which is inconsistent
with the way these things are set (or not set in the case of
rhisf_tbl_0).  The dodgy code is only apparently active when 
we have nn_isf=3 or 2. 

We have to create this branch because we need to debug a region
of code which is affected by numerous changes from different 
GO6 branches, which by some amazing stroke of luck do not 
themselves clash in the revisions used by GO6 but as soon as we 
try to add anything to the buggy region to see what's going on, 
or try to use a later revision of the vn3.6_STABLE_copy branch, 
the fcm merge fails. 

(This is exactly the reason why we need package branches in all 
"standard" configurations and why this needs to be resolved properly at 
source, not on an ad-hoc person by person basis which wastes 
colossal amounts of time and resources.)