Fix to ensure NEMO-CICE continuation runs are bit comparable with NEMO-CICE normal runs, regardless of when the dumps and restarts are perfomed and to ensure things compare over different PE arrangements.
The problems being corrected are data dependent and therefore 
the time of the divergence of models is completely unpredictable, 
often happening several days into a model simulation. 

It's not at all clear that this is an issue only relevant to 
CICE being present. It may well be a general issue for NEMO.