Add changes for compatibility with main MEDUSA branch and to enable coupling of true DMS, CO2 flux, CO2 from atmos and Dust from atmos.
This places responsibility for populating outgoing arrays and 
employing incoming arrays on the MEDUSA branch rather than the 
coupling interface branch. 

NOTE: This revision contains a temporary fix for testing 
to ensure coupling fields are initialised in the absence of
a clear mechanism for obtaining initial start-up values. Remove
the lines labelled RSRH in oce.F90 when this is resolved in the 
MEDUSA code.