Various minor corrections resulting from array bounds checking.
Note: with array bounds check enables, the printing of namelists
natbio and natroam fail because they contain unallocated arrays. 
(This does not cause a problem when not using array bounds checking). 

E.g. arrays such as FRIVER_DEP cause this. It's not immediately
clear how we can correct this without wholesale changes to the 
way these namelists are printed and since it only seems to be
an issue with array bounds checking switched on it may not
be worth pursuing (since we can get round it simply by commenting out 
the offending lines.)