Apply CORRECT intent to fields previously prescribed as OUT.
These should be INOUT fields because they are initialised to 
zero prior to calls to the offending routines. The application 
of OUT effectively renders them undefined unless ALL elements 
are explicitily set in the relevant subroutines, which they 
are not.

It so happens that Cray (and no doubt other compilers) are quite
unfussy about this, but when explicitly useing preset NaNs, things
go badly wrong. Further use of the incorrect intent, while it 
may work in virtually all practical circumstances, is not truly 
portable!